Update files. Three files .c for calculation of statistics - they're starting from compare

+++ source/Tree.h
@@ -17,7 +17,7 @@ typedef struct Node
 {
     char* name; // null if internal
     struct Node** neighbours;
-    unsigned* dist;
+    double* dist;
     Color color;
     char neiNum;
     char maxNeiNum;
@@ -56,10 +56,10 @@ void nodeStackPop(NodeStack* stack);
 Node* nodeCreate(void);
 Node* leafCreate(char* name);
 void nodeDelete(Node* node);
-void nodeAddNeighbour(Node* node, Node* neighbour, unsigned dist);
+void nodeAddNeighbour(Node* node, Node* neighbour, double dist);
 Tree* treeCreate(void);
 void treeDelete(Tree* tree);
-unsigned readLength(char* string, unsigned* pos);
+double readLength(char* string, unsigned* pos);
 char* readName(char* string, unsigned* pos);
 Tree* treeCopy(Tree* source, char copyLCAFinder);
 Tree* treeFromNewick(char* newick);
